Refactored logic into widget
