+++ core-bundle/src/Resources/config/listener.yml
@@ -126,16 +126,6 @@ services:
         arguments:
             - '@database_connection'
 
-    contao.listener.data_container.root_page_dependent_select:
-        class: Contao\CoreBundle\EventListener\DataContainer\RootPageDependentSelectListener
-        arguments:
-            - '@database_connection'
-            - '@translator'
-            - '@contao.routing.scope_matcher'
-            - '@request_stack'
-            - '@contao.csrf.token_manager'
-            - '%contao.csrf_token_name%'
-
     contao.listener.data_container.theme_templates:
         class: Contao\CoreBundle\EventListener\DataContainer\ThemeTemplatesListener
         arguments:
@@ -462,6 +452,14 @@ services:
         class: Contao\CoreBundle\EventListener\Widget\HttpUrlListener
         arguments:
             - '@translator'
+
+    contao.listener.widget.root_page_dependent_select:
+        class: Contao\CoreBundle\EventListener\Widget\RootPageDependentSelectListener
+        arguments:
+            - '@database_connection'
+            - '@translator'
+            - '@contao.csrf.token_manager'
+            - '%contao.csrf_token_name%'
 
     # Backwards compatibility
     Contao\CoreBundle\EventListener\DataContainer\ContentCompositionListener:
